+++ llama_index/chat_engine/utils.py
@@ -1,7 +1,6 @@
 from typing import Optional
 
-from langchain.chat_models.base import BaseChatModel
-from langchain.memory import ChatMessageHistory
+from llama_index.bridge.langchain import BaseChatModel, ChatMessageHistory
 
 from llama_index.chat_engine.types import ChatHistoryType
 from llama_index.indices.service_context import ServiceContext
@@ -12,7 +11,7 @@ def to_chat_buffer(chat_history: ChatHistoryType) -> str:
     buffer = ""
     for human_s, ai_s in chat_history:
         human = "Human: " + human_s
-        ai = "Assistant: " + ai_s
+        ai = "Assistant: " + str(ai_s)
         buffer += "\n" + "\n".join([human, ai])
     return buffer
 
@@ -24,7 +23,7 @@ def to_langchain_chat_history(
     if chat_history is not None:
         for human_message, ai_message in chat_history:
             history.add_user_message(human_message)
-            history.add_ai_message(ai_message)
+            history.add_ai_message(str(ai_message))
     return history