+++ llama_index/indices/struct_store/__init__.py
@@ -1,28 +1,25 @@
 """Structured store indices."""
 
 from llama_index.indices.struct_store.json_query import JSONQueryEngine
-from llama_index.indices.struct_store.pandas import PandasIndex, GPTPandasIndex
-from llama_index.indices.struct_store.pandas_query import (
-    NLPandasQueryEngine,
-    GPTNLPandasQueryEngine,
-)
+from llama_index.indices.struct_store.pandas import GPTPandasIndex, PandasIndex
 from llama_index.indices.struct_store.sql import (
-    SQLStructStoreIndex,
-    SQLContextContainerBuilder,
     GPTSQLStructStoreIndex,
+    SQLContextContainerBuilder,
+    SQLStructStoreIndex,
 )
 from llama_index.indices.struct_store.sql_query import (
-    NLStructStoreQueryEngine,
-    SQLStructStoreQueryEngine,
     GPTNLStructStoreQueryEngine,
     GPTSQLStructStoreQueryEngine,
+    NLStructStoreQueryEngine,
+    SQLStructStoreQueryEngine,
+    SQLTableRetrieverQueryEngine,
+    NLSQLTableQueryEngine,
 )
 
 __all__ = [
     "SQLStructStoreIndex",
     "SQLContextContainerBuilder",
     "PandasIndex",
-    "NLPandasQueryEngine",
     "NLStructStoreQueryEngine",
     "SQLStructStoreQueryEngine",
     "JSONQueryEngine",
@@ -31,5 +28,6 @@ __all__ = [
     "GPTPandasIndex",
     "GPTNLStructStoreQueryEngine",
     "GPTSQLStructStoreQueryEngine",
-    "GPTNLPandasQueryEngine",
+    "SQLTableRetrieverQueryEngine",
+    "NLSQLTableQueryEngine",
 ]