+++ llama_index/objects/__init__.py
@@ -1,0 +1,19 @@
+"""LlamaIndex objects."""
+
+from llama_index.objects.base import ObjectRetriever, ObjectIndex
+from llama_index.objects.base_node_mapping import SimpleObjectNodeMapping
+from llama_index.objects.tool_node_mapping import (
+    SimpleToolNodeMapping,
+    SimpleQueryToolNodeMapping,
+)
+from llama_index.objects.table_node_mapping import SQLTableNodeMapping, SQLTableSchema
+
+__all__ = [
+    "ObjectRetriever",
+    "ObjectIndex",
+    "SimpleObjectNodeMapping",
+    "SimpleToolNodeMapping",
+    "SimpleQueryToolNodeMapping",
+    "SQLTableNodeMapping",
+    "SQLTableSchema",
+]