+++ llama_index/query_engine/__init__.py
@@ -2,6 +2,7 @@ from llama_index.query_engine.citation_query_engine import CitationQueryEngine
 from llama_index.query_engine.flare.base import FLAREInstructQueryEngine
 from llama_index.query_engine.graph_query_engine import ComposableGraphQueryEngine
 from llama_index.query_engine.multistep_query_engine import MultiStepQueryEngine
+from llama_index.query_engine.pandas_query_engine import PandasQueryEngine
 from llama_index.query_engine.retriever_query_engine import RetrieverQueryEngine
 from llama_index.query_engine.retry_query_engine import (
     RetryGuidelineQueryEngine,
@@ -9,6 +10,7 @@ from llama_index.query_engine.retry_query_engine import (
 )
 from llama_index.query_engine.retry_source_query_engine import RetrySourceQueryEngine
 from llama_index.query_engine.router_query_engine import (
+    ToolRetrieverRouterQueryEngine,
     RetrieverRouterQueryEngine,
     RouterQueryEngine,
 )
@@ -25,6 +27,7 @@ __all__ = [
     "MultiStepQueryEngine",
     "RouterQueryEngine",
     "RetrieverRouterQueryEngine",
+    "ToolRetrieverRouterQueryEngine",
     "SubQuestionQueryEngine",
     "SQLJoinQueryEngine",
     "SQLAutoVectorQueryEngine",
@@ -32,4 +35,5 @@ __all__ = [
     "RetrySourceQueryEngine",
     "RetryGuidelineQueryEngine",
     "FLAREInstructQueryEngine",
+    "PandasQueryEngine",
 ]