+++ llama_index/readers/obsidian.py
@@ -9,7 +9,7 @@ import os
 from pathlib import Path
 from typing import Any, List
 
-from langchain.docstore.document import Document as LCDocument
+from llama_index.bridge.langchain import Document as LCDocument
 
 from llama_index.readers.base import BaseReader
 from llama_index.readers.file.markdown_reader import MarkdownReader